+++ server.js
@@ -3,6 +3,7 @@ const fs = require('fs').promises;
 const path = require('path');
 const multer = require('multer');
 const cors = require('cors');
+const { google } = require('googleapis');
 
 const app = express();
 const port = process.env.PORT || 3001;
@@ -13,6 +14,13 @@ app.use(express.json());
 app.use(express.static('build'));
 
 const upload = multer();
+
+// Google Sheets configuration
+const sheets = google.sheets('v4');
+const googleAuth = new google.auth.GoogleAuth({
+  keyFile: process.env.GOOGLE_SERVICE_ACCOUNT_KEY,
+  scopes: ['https://www.googleapis.com/auth/spreadsheets.readonly'],
+});
 
 // File paths
 const DATA_FILE = path.join(__dirname, 'data', 'events.json');
@@ -72,6 +80,36 @@ app.post('/api/sync', upload.none(), async (req, res) => {
   } catch (error) {
     console.error('Error syncing data:', error);
     res.status(500).json({ error: 'Failed to sync data' });
+  }
+});
+
+// Fetch registrants from Google Sheets
+app.post('/api/google-sheets', async (req, res) => {
+  const { spreadsheetId, range } = req.body;
+
+  if (!spreadsheetId || !range) {
+    return res.status(400).json({ error: 'Missing spreadsheetId or range' });
+  }
+
+  try {
+    const authClient = await googleAuth.getClient();
+    const response = await sheets.spreadsheets.values.get({
+      auth: authClient,
+      spreadsheetId,
+      range,
+    });
+
+    const values = response.data.values || [];
+    const registrants = values.map(row => ({
+      name: row[0],
+      email: row[1],
+      phone: row[2],
+    }));
+
+    res.json({ registrants });
+  } catch (error) {
+    console.error('Error fetching data from Google Sheets:', error);
+    res.status(500).json({ error: 'Failed to fetch data from Google Sheets' });
   }
 });
